scripts/requirements.txt
convex
python-dotenv